+++ 03-easy-max-drawdown/go/solution.go
@@ -1,6 +1,6 @@
 package main
 
-func solve(prices []int) int {
+func solve(prices []int64) int64 {
 	_ = prices
 	panic("TODO")
 }
