+++ 04-medium-edge-costs/go/solution.go
@@ -1,6 +1,6 @@
 package main
 
-func solve(n int, edges [][2]int, loads []*int) []int {
+func solve(n int, edges [][2]int, loads []*int64) []int64 {
 	_ = n
 	return nil
 }
