+++ 07-medium-coin-change/go/solution.go
@@ -1,6 +1,6 @@
 package main
 
-func solve(amount int, coins []int) int {
+func solve(amount int, coins []int64) int {
 	_ = amount
 	_ = coins
 	panic("TODO")
