+++ 10-medium-route-costs/go/solution.go
@@ -1,6 +1,6 @@
 package main
 
-func solve(n int, edges [][]int) []int64 {
+func solve(n int, edges [][]int64) []int64 {
 	_ = n
 	_ = edges
 	return nil
