+++ 17-medium-knapsack/go/solution.go
@@ -1,6 +1,6 @@
 package main
 
-func solve(capacity int, items []item) int {
+func solve(capacity int, items []item) int64 {
 	_ = capacity
 	_ = items
 	return 0
