+++ 43-hard-order-book/go/solution.go
@@ -2,7 +2,7 @@ package main
 
 // solve runs the order book and returns
 // [numTrades, p1, q1, p2, q2, ..., bestBid, bidQty, bestAsk, askQty].
-func solve(orders []order) []int {
+func solve(orders []order) []int64 {
 	_ = orders
 	return nil
 }
