+++ 54-hard-spectrum-peptide-recovery/go/solution.go
@@ -1,5 +1,5 @@
 package main
 
-func solve(masses, spectrum []int) ([]int, bool) {
+func solve(masses, spectrum []int64) ([]int64, bool) {
 	return nil, false
 }
